+++ plugins/voxelman/login/plugininfo.d
@@ -1,5 +1,5 @@
-module voxelman.clientdb.plugininfo;
-enum id = "voxelman.clientdb";
+module voxelman.login.plugininfo;
+enum id = "voxelman.login";
 enum semver = "0.6.0";
 enum deps = [];
 enum clientdeps = [];